Combine AdminAppDataTestBase and AdminAppDbContextTestBase now that the hierarchy provides no assistance.
