+++ Application/EdFi.Ods.AdminApp.Management.Core.Tests/EdFi.Ods.AdminApp.Management.Core.Tests.csproj
@@ -48,7 +48,6 @@
         <!-- <Compile Include="../EdFi.Ods.AdminApp.Management.Tests/User/GetAdminAppUsersQueryTests.cs" LinkBase="User" /> -->
         <!-- <Compile Include="../EdFi.Ods.AdminApp.Management.Tests/User/AddUserTests.cs" LinkBase="User" /> -->
         <Compile Include="../EdFi.Ods.AdminApp.Management.Tests/AdminAppDataTestBase.cs" />
-        <Compile Include="../EdFi.Ods.AdminApp.Management.Tests/AdminAppDbContextTestBase.cs" />
         <!-- <Compile Include="../EdFi.Ods.AdminApp.Management.Tests/AutoMapper/AutoMapperTests.cs" LinkBase="AutoMapper" /> -->
         <Compile Include="../EdFi.Ods.AdminApp.Management.Tests/ClaimSetEditor/ClaimSetFileImportCommandTests.cs" LinkBase="ClaimSetEditor" />
         <Compile Include="../EdFi.Ods.AdminApp.Management.Tests/ClaimSetEditor/DeleteResourceOnClaimSetCommandTests.cs" LinkBase="ClaimSetEditor" />
